Fix indentation errors in the code and add a module.scss document to each component
