+++ src/components/footer/footer.module.scss
@@ -1,0 +1,13 @@
+@import "../../css/colors.scss";
+
+.container {
+  width: 100%;
+  justify-content: center;
+  display: flex;
+  position: relative;
+  overflow: hidden;
+  height: 40vh;
+  background-color: $footer-container;
+  flex-direction: column;
+  align-items: center;
+}
