+++ src/components/footer/footer.tsx
@@ -1,0 +1,12 @@
+import * as React from "react";
+
+// import styled from 'styled-components'; 
+import * as styledFooter from "./footer.module.scss";  
+
+const Footer = ({children}) => (
+  <div className={styledFooter.container}>
+    {children}
+  </div>
+);
+
+export default Footer;
