+++ src/components/header/header.module.scss
@@ -1,0 +1,4 @@
+.container {
+  display: flex;
+  justify-content: space-between;
+}
