+++ src/components/mdx-helpers/git-growth-trend.js
@@ -1,4 +1,3 @@
-
 import React from "react"
 import Helmet from "react-helmet"
