+++ src/components/nav-bar/nav-bar.module.scss
@@ -1,0 +1,6 @@
+.container {
+  grid-column-start: 2;
+  grid-column-end: span 11;
+  grid-row-start: 1;
+  grid-row-end: span 1;
+}
