+++ src/components/nav-menu/nav-menu.module.scss
@@ -1,0 +1,10 @@
+@import "../../css/colors.scss";
+
+.linkMenu {
+  margin-top: 10px;
+  margin-bottom: 10px;
+  color: $black;
+  font-size: 16px;
+  line-height: 20px;
+  font-weight: 700;
+}
