+++ src/components/pager/pager.module.scss
@@ -1,4 +1,4 @@
-@import '../css/colors.scss';
+@import "../../css/colors.scss";
 
 .pager-container {
   display: flex;
