+++ src/components/pager/pager.tsx
@@ -1,60 +1,62 @@
-import React from "react"
-import * as pagerStyles from "../css/pager.module.scss"
-import Pages from "../components/pages"
+import React from "react";
 
-import { Link } from "gatsby"
+import * as pagerStyles from "./pager.module.scss";
+
+import Pages from "../pages/pages";
+
+import { Link } from "gatsby";
 
 const setPrevPage = ({currentPage}) => {
   if (currentPage === 1) {
-    return 'javascript:void(0)'
+  return 'javascript:void(0)'
   }
   
   if ((currentPage - 1) === 1) {
-    return '/'
+  return '/'
   }
 
   return `/page/${(currentPage - 1)}`
-}
+};
 
 const setNextPage = ({numPages, currentPage}) => {
   if (currentPage === numPages) {
-    return 'javascript:void(0)'
+  return 'javascript:void(0)'
   }
   return `/page/${(currentPage + 1)}`
-}
+};
 
 const setPagesData = ({numPages, currentPage}) => {
   if (currentPage + 2 >= numPages) {
-    return {
-      firstPage: (numPages - 4),
-      secondPage: (numPages - 3),
-      thirdPage: (numPages - 2),
-      fourthPage: (numPages - 1),
-      lastPage: numPages,
-      currentPage: currentPage
-    };
+  return {
+    firstPage: (numPages - 4),
+    secondPage: (numPages - 3),
+    thirdPage: (numPages - 2),
+    fourthPage: (numPages - 1),
+    lastPage: numPages,
+    currentPage: currentPage
+  };
   } 
   
   if (currentPage - 2 <= 1) {
-    return {
-      firstPage: 1,
-      secondPage: 2,
-      thirdPage: 3,
-      fourthPage: 4,
-      lastPage: 5,
-      currentPage: currentPage
-    };
+  return {
+    firstPage: 1,
+    secondPage: 2,
+    thirdPage: 3,
+    fourthPage: 4,
+    lastPage: 5,
+    currentPage: currentPage
+  };
   }
 
   return {
-    firstPage: (currentPage - 2),
-    secondPage: (currentPage - 1),
-    thirdPage: currentPage,
-    fourthPage: (currentPage + 1),
-    lastPage: (currentPage + 2),
-    currentPage: currentPage
+  firstPage: (currentPage - 2),
+  secondPage: (currentPage - 1),
+  thirdPage: currentPage,
+  fourthPage: (currentPage + 1),
+  lastPage: (currentPage + 2),
+  currentPage: currentPage
   }
-}
+};
 
 const Pager = ({ pageContext }) => {
   const {numPages,  currentPage} = pageContext
@@ -64,14 +66,14 @@ const Pager = ({ pageContext }) => {
   return(
     <div className={pagerStyles.pagerContainer}>
       <Link className={currentPage === 1 ? pagerStyles.disabledPagerLink : pagerStyles.pagerLink} to={prevPage} rel="prev">
-      ← Prev
+        ← Prev
       </Link>
       <Pages data={pagesData}/>
       <Link className={currentPage === numPages ? pagerStyles.disabledPagerLink : pagerStyles.pagerLink} to={nextPage} rel="next">
-      Next →
+        Next →
       </Link>
     </div>
-  )
-}
+  );
+};
 
-export default Pager
+export default Pager;
