+++ src/components/profile/profile.tsx
@@ -1,7 +1,7 @@
 import * as React from "react"
 
-const Profile = () => {
-    return <h1>Profile</h1>
-}
+const Profile = () => (
+  <h1>Profile</h1>
+);
 
 export default Profile;
