+++ src/css/colors.scss
@@ -1,4 +1,13 @@
+// Pink
 $xl-pink: #EE1A64;
+
+// Yellow
 $default-yellow: #F6C14D;
+
+// Gray
 $default-gray: #858585;
 $disabled-gray: #B6B6B6;
+
+// Dark
+$black: #000000;
+$footer-container: #161616;
