+++ src/layouts/layout.module.scss
@@ -1,6 +1,6 @@
 @import "../css/typography.css";
 
-body{
+body {
     font-family: 'gilroy', serif;
 }
